python_CV_Parser/requirements.txt
numpy==2.2.4
opencv-python>=4.11
icecream>=2.1.4 
pyserial>=3.4